Deselect columns automatically, created viewmode slot on rdt
