+++ src/components/RoarDataTable.vue
@@ -4,6 +4,11 @@
   </div>
   <div v-else>
     <div class="w-full gap-2 pt-4 flex justify-content-center flex-wrap">
+      <span>
+        <div class="relative">
+          <slot />
+        </div>
+      </span>
       <span class="p-float-label">
         <PvMultiSelect
           id="ms-columns"
@@ -31,21 +36,6 @@
           @update:model-value="onFreezeToggle"
         />
         <label for="ms-columns">Freeze Columns</label>
-      </span>
-      <span>
-        <div class="relative">
-          <!-- <span>View</span> -->
-          <label for="ms-columns" class="view-label">View</label>
-          <PvDropdown
-            id="view-columns"
-            v-model="viewMode"
-            :options="viewOptions"
-            option-label="label"
-            option-value="value"
-            class="ml-2"
-          />
-          
-        </div>
       </span>
       <span v-if="allowExport" class="flex flex-row flex-wrap justify-content-end">
         <PvButton label="Export Selected"  :disabled="selectedRows.length === 0" @click="exportCSV(true, $event)"  />
@@ -235,11 +225,14 @@ import _get from 'lodash/get';
 import _set from 'lodash/set';
 import _map from 'lodash/map';
 import _forEach from 'lodash/forEach';
+import _debounce from 'lodash/debounce';
 import _find from 'lodash/find';
 import _filter from 'lodash/filter';
 import _toUpper from 'lodash/toUpper';
 import _startCase from 'lodash/startCase';
+import _without from 'lodash/without';
 import InputSwitch from 'primevue/inputswitch';
+import { property } from 'lodash';
 // import Checkbox from 'primevue/checkbox';
 
 /*
@@ -280,17 +273,17 @@ const props = defineProps({
   lazy: { type: Boolean, default: false },
 });
 
-const viewOptions = ref([
-  { label: 'Color', value: 'color' },
-  { label: 'Percentile', value: 'percentile' },
-  { label: 'Standard Score', value: 'standard' },
-  { label: 'Raw Score', value: 'raw' },
-]);
-
-const viewMode =ref(false);
-
 const inputColumns = ref(props.columns);
-const selectedColumns = ref(props.columns);
+const selectedColumns = ref(
+  _without(props.columns.map((col) => {
+    if(col.dataType === 'score') return col
+    for (const row of props.data){
+      if(_get(row, col.field)){
+        return col;
+      }
+    }
+  }), undefined)
+)
 // Filter the live data (props.columns) with the selections of selectedColumns
 const computedColumns = computed(() => {
   return _map(selectedColumns.value, (col) => {
